Переименование файлов и компонент
