+++ src/pages/login-page/login-page.tsx
@@ -1,4 +1,4 @@
-function LoginScreen(): JSX.Element {
+function LoginPage(): JSX.Element {
 
   return (
     <div className="page page--gray page--login">
@@ -62,4 +62,4 @@ function LoginScreen(): JSX.Element {
   );
 }
 
-export default LoginScreen;
+export default LoginPage;
